Add log handling for apps
