+++ src/public/publicAPIs.ts
@@ -5,7 +5,7 @@ import { ExtendedWindow, MessageEvent } from "../internal/interfaces";
 import { settings } from "./settings";
 import { TabInformation, TabInstanceParameters, TabInstance, DeepLinkParameters, Context } from "./interfaces";
 import { getGenericOnCompleteHandler } from "../internal/utils";
-import { registerGetLogHandler } from "../private/privateAPIs";
+import { files } from "../private/files";
 
 // ::::::::::::::::::::::: MicrosoftTeams SDK public API ::::::::::::::::::::
 /**
@@ -59,7 +59,7 @@ export function initialize(hostWindow: any = window): void {
         registerFullScreenHandler(null);
         registerBackButtonHandler(null);
         registerBeforeUnloadHandler(null);
-        registerGetLogHandler(null);
+        files.registerGetLogHandler(null);
       }
 
       if (GlobalVars.frameContext === frameContexts.settings) {
